feat: change TalkToListenRatio fields to float64 for more precise percentage representation

+++ internal/modules/analysis/model.go
@@ -18,8 +18,8 @@ type RoleMapping struct {
 
 // TalkToListenRatio represents the proportion of speech time between manager and client.
 type TalkToListenRatio struct {
-	ManagerPercentage int `json:"manager_percentage"`
-	ClientPercentage  int `json:"client_percentage"`
+	ManagerPercentage float64 `json:"manager_percentage"`
+	ClientPercentage  float64 `json:"client_percentage"`
 }
 
 // AwkwardPause records dead air or prolonged hesitation in seconds.
